DNS/FLOW ML refactor.
