+++ src/main/scala/Dispatcher.scala
@@ -9,8 +9,8 @@ object Dispatcher {
       val f = args(0)
 
       f match {
-        case "dns_post_lda" => dns_post_lda.run()
-        case "dns_pre_lda" => dns_pre_lda.run()
+        case "dns_post_lda" => DnsPostLDA.run()
+        case "dns_pre_lda" => DnsPreLDA.run()
         case "flow_post_lda" =>FlowPostLDA.run()
         case "flow_pre_lda" => FlowPreLDA.run()
       }
